* 'master' of git://scm.dev.nokia.troll.no/qt/qt-fire-team: (116 commits)
  Update Japanese translations.
  Updated Russian translation
  Update Korean translations to 4.8
  qmake: fix writing of incremental linker option for vcxproj files
  support the obj-c #import statement in the dependency generator
  runonphone timestamps nanosecond formating set to 9 chars, 0 padded
  Put Qt translation files to rom on Symbian.
  Add -nn for .cpp MOC preprocessing for INTEGRITY
  Remove commented out code.
  Update Polish translations for 4.8
  Fix QDeclarativeEngine::setOfflineStoragePath() for Symbian
  Fixes: libpng symbols exported from QtGui.dll on Windows
  Converting from double to qreal in gui
  Doc updates to installation, platform notes and symbian introduction.
  Fix XML schema validation failure.
  purge "Translation Rules for Plurals"
  add -list-languages option to lupdate
  Fix construction races in QtNetwork
  Removing accidental use of double instead of qreal from QLine
  Fix incorrect QFileInfo permissions on windows
  ...

+++ src/declarative/util/qdeclarativeanimation.cpp
@@ -2719,7 +2719,7 @@ void QDeclarativeParentAnimation::transition(QDeclarativeStateActions &actions,
                     }
 
                     if (scale != 0)
-                        rotation = atan2(transform.m12()/scale, transform.m11()/scale) * 180/M_PI;
+                        rotation = atan2(transform.m12()/scale, transform.m11()/scale) * 180/qreal(M_PI);
                     else {
                         qmlInfo(this) << QDeclarativeParentAnimation::tr("Unable to preserve appearance under scale of 0");
                         ok = false;
